# NMS with CUDA only