--- a/mmcv/ops/csrc/pytorch/cuda/nms_cuda.cu
+++ b/mmcv/ops/csrc/pytorch/cuda/nms_cuda.cu
@@ -24,31 +24,41 @@ Tensor NMSCUDAKernelLauncher(Tensor boxes, Tensor scores, float iou_threshold,
       boxes_num, iou_threshold, offset, boxes_sorted.data_ptr<float>(),
       (unsigned long long*)mask.data_ptr<int64_t>());
 
-  at::Tensor mask_cpu = mask.to(at::kCPU);
-  unsigned long long* mask_host =
-      (unsigned long long*)mask_cpu.data_ptr<int64_t>();
-
-  std::vector<unsigned long long> remv(col_blocks);
-  memset(&remv[0], 0, sizeof(unsigned long long) * col_blocks);
-
-  at::Tensor keep_t =
-      at::zeros({boxes_num}, boxes.options().dtype(at::kBool).device(at::kCPU));
-  bool* keep = keep_t.data_ptr<bool>();
-
-  for (int i = 0; i < boxes_num; i++) {
-    int nblock = i / threadsPerBlock;
-    int inblock = i % threadsPerBlock;
-
-    if (!(remv[nblock] & (1ULL << inblock))) {
-      keep[i] = true;
-      // set every overlap box with bit 1 in remv
-      unsigned long long* p = mask_host + i * col_blocks;
-      for (int j = nblock; j < col_blocks; j++) {
-        remv[j] |= p[j];
-      }
-    }
-  }
-
+  at::Tensor keep_t = at::zeros(
+      {boxes_num}, boxes.options().dtype(at::kBool).device(at::kCUDA));
+  gather_keep_from_mask_parallize<<<
+      1, THREADS_PER_BLOCK, col_blocks * sizeof(unsigned long long), stream>>>(
+      keep_t.data_ptr<bool>(), (unsigned long long*)mask.data_ptr<int64_t>(),
+      boxes_num);
   AT_CUDA_CHECK(cudaGetLastError());
-  return order_t.masked_select(keep_t.to(at::kCUDA));
+  return order_t.masked_select(keep_t);
+
+  // at::Tensor mask_cpu = mask.to(at::kCPU);
+  // unsigned long long* mask_host =
+  //     (unsigned long long*)mask_cpu.data_ptr<int64_t>();
+
+  // std::vector<unsigned long long> remv(col_blocks);
+  // memset(&remv[0], 0, sizeof(unsigned long long) * col_blocks);
+
+  // at::Tensor keep_t =
+  //     at::zeros({boxes_num},
+  //     boxes.options().dtype(at::kBool).device(at::kCPU));
+  // bool* keep = keep_t.data_ptr<bool>();
+
+  // for (int i = 0; i < boxes_num; i++) {
+  //   int nblock = i / threadsPerBlock;
+  //   int inblock = i % threadsPerBlock;
+
+  //   if (!(remv[nblock] & (1ULL << inblock))) {
+  //     keep[i] = true;
+  //     // set every overlap box with bit 1 in remv
+  //     unsigned long long* p = mask_host + i * col_blocks;
+  //     for (int j = nblock; j < col_blocks; j++) {
+  //       remv[j] |= p[j];
+  //     }
+  //   }
+  // }
+
+  // AT_CUDA_CHECK(cudaGetLastError());
+  // return order_t.masked_select(keep_t.to(at::kCUDA));
 }