--- a/mmcv/ops/csrc/pytorch/cuda/nms_cuda.cu
+++ b/mmcv/ops/csrc/pytorch/cuda/nms_cuda.cu
@@ -26,8 +26,9 @@ Tensor NMSCUDAKernelLauncher(Tensor boxes, Tensor scores, float iou_threshold,
 
   at::Tensor keep_t = at::zeros(
       {boxes_num}, boxes.options().dtype(at::kBool).device(at::kCUDA));
-  gather_keep_from_mask_parallize<<<
-      1, THREADS_PER_BLOCK, col_blocks * sizeof(unsigned long long), stream>>>(
+  gather_keep_from_mask_parallize<<<1, min(col_blocks, THREADS_PER_BLOCK),
+                                    col_blocks * sizeof(unsigned long long),
+                                    stream>>>(
       keep_t.data_ptr<bool>(), (unsigned long long*)mask.data_ptr<int64_t>(),
       boxes_num);
   AT_CUDA_CHECK(cudaGetLastError());